Resolves routing errors, renames all page names containing 'Results' to 'Result' for consistency
