+++ client/containers/ResultPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import ResultPageEntry from '../pages/Results/ResultsPageEntry.js';
+import ResultPageEntry from '../pages/Result/ResultPageEntry.js';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import resultSelector from '../actions/index';
@@ -14,7 +14,7 @@ class ResultPage extends React.Component {
     return (
       <div>
         {this.props.moods.map((mood, i) =>
-          <MoodPageEntry
+          <ResultPageEntry
             onClick={this.onMoodClick}
             key={i}
             moods={mood} />
@@ -29,15 +29,15 @@ class ResultPage extends React.Component {
 //moodData will now be available as props in MoodPage class
 function mapStateToProps(state) {
   return {
-    moods: state.moods
+    results: state.results
   }
 }
 
 
 function mapDispatchToProps(dispatch) {
-  return bindActionCreators({moodSelector: moodSelector}, dispatch);
+  return bindActionCreators({resultSelector: resultSelector}, dispatch);
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(ResultsPage);
+export default connect(mapStateToProps, mapDispatchToProps)(ResultPage);
