+++ client/pages/Result/ResultPageEntry.js
